+++ key_result/application/interfaces.py
@@ -1,12 +1,21 @@
 from abc import ABC, abstractmethod
 from typing import List, Optional
 
-from key_result.domain.entities import KeyResult
+from key_result.domain.entities.key_result import KeyResult
+from key_result.domain.entities.key_result_with_tasks import KeyResultWTasks
 
 
 class IKeyResultApplicationService(ABC):
     @abstractmethod
     def get_key_results_by_team_id(self, team_id, filters) -> Optional[List[KeyResult]]:
+        pass
+    
+    @abstractmethod
+    def get_key_results_by_user_id(self, user_id) -> Optional[List[KeyResult]]:
+        pass
+    
+    @abstractmethod
+    def get_key_results_and_tasks_by_user_id(self, user_id) -> Optional[List[KeyResultWTasks]]:
         pass
     
     @abstractmethod
@@ -21,6 +30,14 @@ class IKeyResultApplicationService(ABC):
 class IKeyResultRepository(ABC):
     @abstractmethod
     def find_by_team_id(self, team_id, filters) -> Optional[List[KeyResult]]:
+        pass
+    
+    @abstractmethod
+    def find_by_user_id(self, user_id) -> Optional[List[KeyResult]]:
+        pass
+    
+    @abstractmethod
+    def find_kr_and_tasks_by_user_id(self, user_id: str) -> Optional[List[KeyResultWTasks]]:
         pass
     
     @abstractmethod
