+++ key_result/domain/entities/key_result_check_mark.py
@@ -1,0 +1,15 @@
+from datetime import datetime
+from pydantic import BaseModel
+from typing import Optional
+
+
+class KeyResultCheckMark(BaseModel):
+    state: str
+    description: str
+    key_result: str
+    user: str
+    assigned_user: Optional[str]
+    id: Optional[str]
+    created_at: Optional[datetime]
+    updated_at: Optional[datetime]
+    deleted_at: Optional[datetime]
