+++ key_result/domain/entities/key_result_support_team_members_user.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from typing import Optional
+
+
+class KeyResultSupportTeamMembersUser(BaseModel):
+    key_result: Optional[str]
+    user: Optional[str]
