+++ key_result/infrastructure/api/v1/endpoints.py
@@ -4,7 +4,9 @@ from ninja import Query, Router
 from key_result.application.interfaces import IKeyResultApplicationService
 from key_result.application.services import KeyResultApplicationService
 
-from key_result.domain.entities import KeyResult
+from key_result.domain.entities.key_result import KeyResult
+from key_result.domain.entities.key_result_with_tasks import KeyResultWTasks
+
 from key_result.infrastructure.api.v1.filters import KeyResultFilterSchema
 
 kr_router = Router(tags=["key_result"])
@@ -19,6 +21,16 @@ def get_kr(request, key_result_id: str):
 @kr_router.get("/team/{team_id}", response={200: List[KeyResult], 404: dict})
 def get_kr_by_team(request, team_id: str, filters: KeyResultFilterSchema = Query(...)):
     data = kr_service.get_key_results_by_team_id(team_id, filters)
+    return data
+
+@kr_router.get("/user/{user_id}", response={200: List[KeyResult], 404: dict})
+def get_kr_by_owner(request, user_id: str):
+    data = kr_service.get_key_results_by_user_id(user_id)
+    return data
+
+@kr_router.get("/tasks/user/{user_id}", response={200: List[KeyResultWTasks], 404: dict})
+def get_key_result_tasks_by_owner(request, user_id: str):
+    data = kr_service.get_key_results_and_tasks_by_user_id(user_id)
     return data
 
 @kr_router.patch("/{key_result_id}", response={200: KeyResult, 404: dict})
