+++ task_manager/domain/entities/task_history.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+from typing import Optional
+
+
+class TaskHistory(BaseModel):
+    task: str
+    field: str
+    old_state: Optional[str]
+    new_state: Optional[str]
+    author: Optional[str]
